Gather events into a separate module
