+++ cast/app.py
@@ -1,31 +1,15 @@
 #!/usr/bin/env python3
-import argparse
-
-from flask import (
-    Flask,
-    render_template,
-    request,
-    send_from_directory,
-    redirect,
-    jsonify,
-)
-
-import sys
-
-import flask_socketio
-from flask_jwt_extended import JWTManager
-
-
-import json
-from .routes import http
-from .logger import Logging
-import pty
 import os
-import subprocess
+import sys
 import select
 import shlex
 import random
 import string
+import argparse
+
+from flask import Flask
+import flask_socketio
+from flask_jwt_extended import JWTManager
 
 __version__ = "0.0.1"
 
@@ -39,8 +23,6 @@ app.config["sessions"] = {}
 app.config["log_file"] = r"log_data/"
 app.config["JWT_TOKEN_LOCATION"] = ["cookies"]
 
-app.register_blueprint(http, url_prefix=r"")
-
 
 def random_string(string_length=10):
     """Generate a random string of fixed length """
@@ -52,6 +34,10 @@ app.config["JWT_SECRET_KEY"] = random_string()
 
 jwt = JWTManager(app)
 socketio = flask_socketio.SocketIO(app)
+
+from .routes import http
+
+app.register_blueprint(http, url_prefix=r"")
 
 
 def read_and_forward_pty_output(session_id):
@@ -83,131 +69,6 @@ def read_and_forward_pty_output(session_id):
                     except OSError:
                         socketio.emit("disconnect", namespace="/cast")
                         sys.exit(0)
-
-
-@jwt.unauthorized_loader
-def unauthorized_response(callback):
-    return redirect("/")
-
-
-@jwt.invalid_token_loader
-def invalid_token(callback):
-    return redirect("/")
-
-
-@socketio.on("new-session", namespace="/cast")
-def new_session(data=None):
-    """To register session on WebSocket server
-    Similar to 'connect()', used for adding a new tab session
-    """
-    session_id = ""
-    if data is not None:
-        session_id = data["session_id"]
-    print("new-session: {}\n".format(session_id))
-
-    if session_id in app.config["sessions"]:
-        return
-
-    (child_pid, fd) = pty.fork()
-
-    if child_pid == 0:
-        # child: run system command
-        subprocess.run(app.config["cmd"])
-    else:
-        # parent: stream input/output from child to Socketio web channel
-        app.config["sessions"][session_id] = {}
-        app.config["sessions"][session_id]["fd"] = fd
-        app.config["sessions"][session_id]["child_pid"] = child_pid
-
-        cmd = " ".join(shlex.quote(c) for c in app.config["cmd"])
-
-        print("new-session: child pid is", child_pid)
-        print(
-            "new-session: starting background task with command `{}` to continously read "
-            "and forward pty output to client".format(cmd)
-        )
-        """
-        print(
-            f"new-session: starting background task with command `{cmd}` to continously read "
-            "and forward pty output to client"
-        )
-        """
-
-        socketio.start_background_task(
-            target=read_and_forward_pty_output, session_id=session_id
-        )
-
-        print("new-session: task started")
-
-
-@socketio.on("connect", namespace="/cast")
-def connect(data=None):
-    session_id = (
-        request.values.get("session_id")
-        if not request.values.get("session_id") is None
-        else ""
-    )
-    if session_id == "" and data is not None:
-        session_id = data["session_id"]
-        print("connect: {}\n".format(session_id))
-
-    # Create new session only when id not in records
-    if session_id in app.config["sessions"]:
-        return
-
-    (child_pid, fd) = pty.fork()
-
-    if child_pid == 0:
-        # child: start system command
-        subprocess.run(app.config["cmd"])
-    else:
-        # parent: print info, stream child input/output to socketio
-        # Store sessions by ssid
-        print("opening a new session")
-        app.config["sessions"] = {}
-        app.config["sessions"][session_id] = {}
-        app.config["sessions"][session_id]["fd"] = fd
-        app.config["sessions"][session_id]["child_pid"] = child_pid
-
-        cmd = " ".join(shlex.quote(c) for c in app.config["cmd"])
-
-        print("connect: child pid is", child_pid)
-        print(
-            "connect: starting background task with command `{}` to continously read "
-            "and forward pty output to client".format(cmd)
-        )
-        """
-        print(
-            f"new-session: starting background task with command `{cmd}` to continously read "
-            "and forward pty output to client"
-        )
-        """
-
-        # Output terminal message corresponding to ssid
-        socketio.start_background_task(
-            target=read_and_forward_pty_output, session_id=session_id
-        )
-
-        print("connect: task started")
-
-
-@socketio.on("client-input", namespace="/cast")
-def client_input(data):
-    # Update current session
-    app.config["current_session"] = data["session_id"]
-    print("input: {}".format(app.config["sessions"]))
-    log = Logging(app.config["current_session"])
-
-    if data["session_id"] in app.config["sessions"]:
-        file_desc = app.config["sessions"][data["session_id"]]["fd"]
-
-        if file_desc:
-            if data["input"] == "":
-                # When switching sessions, send a key to update terminal content
-                os.write(file_desc, b"\x00")
-            else:
-                log.write_log(data["input"])
-                os.write(file_desc, data["input"].encode())
 
 
 def create_parser():
